Update Dockerfile and remove numpy from requirements.txt
This update streamlines the Dockerfile by adding a stage for packaging Python code and copying necessary files from the builder. It also removes numpy from the requirements.txt file as it's included in the Dockerfile separately. This helps to reduce redundancy and improve the efficiency of the Docker build process.